Move deploy_util.go -> testutil/deploy.go
